refactor(server): rename resolveGroupPath module to GroupPath
Renames the module after its data structure (group path) rather than
its primary operation, and narrows the public surface to just
resolveGroupPathOrDie since resolveGroupPath itself was only referenced
by its own test.
