+++ packages/server/src/GroupImpl.ts
@@ -14,7 +14,7 @@ import {
 } from "effect";
 import type * as Api from "./Api";
 import type * as FunctionImpl from "./FunctionImpl";
-import { resolveGroupPathOrDie } from "./resolveGroupPath";
+import { resolveGroupPathOrDie } from "./GroupPath";
 
 export const TypeId = "@confect/server/GroupImpl";
 export type TypeId = typeof TypeId;
